+++ account/tests/test_urls.py
@@ -1,0 +1,1 @@
+"""Python module for writting tests for urls"""