+++ CHANGELOG.md
@@ -8,3 +8,10 @@
 ## 2020-01-22
 * added flake8 test
 * created test.sh script for runing tests
+
+## 2020-01-24
+* added login functionality
+* added logout functionality
+* added test for login functionality
+* added test for logout functionality
+* updated some documentations